Crate taskqueue1_beta2

Features

Handle the following Resources with ease from the central hub ...

taskqueues (get)

tasks (delete, get, insert, lease, list, patch and update)

Everything else about the taskqueue API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.tasks().lease(...).doit()
let r = hub.tasks().insert(...).doit()
let r = hub.tasks().delete(...).doit()
let r = hub.tasks().patch(...).doit()
let r = hub.tasks().list(...).doit()
let r = hub.tasks().get(...).doit()
let r = hub.tasks().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
taskqueue1_beta2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate taskqueue1_beta2;
use taskqueue1_beta2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use taskqueue1_beta2::Taskqueue;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Taskqueue::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.tasks().lease("project", "taskqueue", -8, -70)
             .tag("amet.")
             .group_by_tag(false)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Task
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

lease tasks (none)

insert tasks (request|response)

delete tasks (none)

patch tasks (request|response)

list tasks (none)

get tasks (response)

update tasks (request|response)

TaskDeleteMethodBuilder
Delete a task from a TaskQueue.

TaskGetMethodBuilder
Get a particular task from a TaskQueue.

TaskInsertMethodBuilder
Insert a new task in a TaskQueue

TaskLeaseMethodBuilder
Lease 1 or more tasks from a TaskQueue.

TaskListMethodBuilder
List Tasks in a TaskQueue

TaskMethodsBuilder
A builder providing access to all methods supported on task resources. It is not used directly, but through the Taskqueue hub.

TaskPatchMethodBuilder
Update tasks that are leased out of a TaskQueue. This method supports patch semantics.

TaskQueue
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get taskqueues (response)

TaskQueueAcl
ACLs that are applicable to this TaskQueue object.

This type is not used in any activity, and only used as part of another schema.

TaskQueueStats
Statistics for the TaskQueue object in question.

This type is not used in any activity, and only used as part of another schema.

TaskUpdateMethodBuilder
Update tasks that are leased out of a TaskQueue.

Taskqueue
Central instance to access all Taskqueue related resource activities

TaskqueueGetMethodBuilder
Get detailed information about a TaskQueue.

TaskqueueMethodsBuilder
A builder providing access to all methods supported on taskqueue resources. It is not used directly, but through the Taskqueue hub.

Tasks
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

lease tasks (response)

Tasks2
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tasks (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.